Refac : Common Components styled 분리
