--- src/_shared/componentProps.js
@@ -1,9 +1,0 @@
-export const THEME = {
-  DARK: "dark",
-  LIGHT: "light",
-};
-
-export const TYPE = {
-  DEFAULT: "default",
-  TRANSPARENT: "transparent",
-};
