+++ src/components/commons/Badge/Badge.jsx
@@ -1,12 +1,11 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
-import { colors, fontWeight, fontSize, loadings } from "@/_shared";
+import { Layout, Text } from "./Badge.styled";
 
 const THEME = {
-  LIGHT: "light",
   DARK: "dark",
+  LIGHT: "light",
 };
 
 const MODE = {
@@ -39,54 +38,5 @@ Badge.defaultProps = {
   isBold: false,
   isLoading: false,
 };
-
-const textColor = {
-  light: {
-    primary: colors.white,
-    secondary: colors.gray600,
-  },
-  dark: {
-    primary: colors.white,
-    secondary: colors.white,
-  },
-};
-
-const bgColor = {
-  light: {
-    primary: colors.blue100,
-    secondary: colors.gray100,
-  },
-  dark: {
-    primary: colors.blue100,
-    secondary: colors.gray700,
-  },
-};
-
-const Layout = styled.div`
-  display: inline-flex;
-  align-items: center;
-
-  border-radius: 8px;
-  background: ${(props) => bgColor[props.theme][props.mode]};
-
-  ${(props) => props.isLoading && loadings[props.theme]};
-
-  ${(props) =>
-    props.isLoading &&
-    `
-    min-width: 48px;
-    height: 20px;
-    `}
-`;
-
-const Text = styled.span`
-  padding: 4px 8px;
-
-  color: ${(props) => textColor[props.theme][props.mode]};
-  font-size: ${fontSize.xxs};
-  font-weight: ${(props) =>
-    props.isBold ? fontWeight.bold : fontWeight.regular};
-  line-height: 12px;
-`;
 
 export default Badge;
