+++ src/components/commons/ImageBox/ImageBox.jsx
@@ -1,9 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
+import { ImageBoxWrapper } from "./ImageBox.styled";
 import { Logo } from "@/foundations";
-import { colors, loadings } from "@/_shared";
 
 const THEME = {
   LIGHT: "light",
@@ -59,56 +58,3 @@ ImageBox.defaultProps = {
 };
 
 export default ImageBox;
-
-const imageBgColor = {
-  light: colors.gray25,
-  dark: colors.gray900,
-};
-
-const imageWidth = {
-  large: "400px",
-  medium: "300px",
-  small: "220px",
-};
-
-const imageHeight = {
-  large: "200px",
-  medium: "150px",
-  small: "110px",
-};
-
-const borderRadius = {
-  large: "12px",
-  medium: "8px",
-  small: "8px",
-};
-
-const ImageBoxWrapper = styled.div`
-  display: flex;
-  width: ${(props) => imageWidth[props.size]};
-  height: ${(props) => imageHeight[props.size]};
-  border-radius: ${(props) => borderRadius[props.size]};
-
-  background: ${(props) => imageBgColor[props.theme]};
-
-  align-items: center;
-  justify-content: center;
-
-  ${(props) => props.isLoading && loadings[props.theme]};
-
-  overflow: hidden;
-  transition: 0.3s;
-
-  svg {
-    width: 33%;
-    height: 33%;
-    filter: ${(props) => props.isDisabled && `blur(5px)`};
-  }
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    filter: ${(props) => props.isDisabled && `blur(5px)`};
-  }
-`;
