+++ src/components/commons/InputImage/InputImage.jsx
@@ -1,11 +1,10 @@
 import React, { forwardRef } from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
 import Dropzone from "react-dropzone";
 
+import { Layout, Title, Description } from "./InputImage.styled";
 import { Icon } from "@/foundations";
-import { colors, fontSize, fontWeight, shadows } from "@/_shared";
 
 const THEME = {
   LIGHT: "light",
@@ -50,62 +49,3 @@ InputImage.defaultProps = {
 };
 
 export default InputImage;
-
-const bgColor = {
-  light: colors.white,
-  dark: colors.gray900,
-};
-
-// const borderColor = {
-//   light: colors.gray300,
-//   dark: colors.gray700,
-// };
-
-const borderColor = {
-  light: {
-    default: colors.gray300,
-    error: colors.errorOpacity200,
-  },
-  dark: {
-    default: colors.gray700,
-    error: colors.errorOpacity200,
-  },
-};
-
-const iconColor = {
-  light: colors.gray300,
-  dark: colors.gray600,
-};
-
-const Layout = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5rem;
-
-  padding: 3rem;
-  border-radius: 8px;
-  border: 2px dashed ${(props) => borderColor[props.theme][props.status]};
-  box-shadow: ${shadows.button};
-
-  background-color: ${(props) => bgColor[props.theme]};
-  cursor: pointer;
-
-  svg {
-    width: 42px;
-    height: 42px;
-    stroke: ${(props) => iconColor[props.theme]};
-  }
-`;
-
-const Title = styled.div`
-  padding-top: 1rem;
-  color: ${colors.blue100};
-  font-weight: ${fontWeight.bold};
-`;
-
-const Description = styled.div`
-  color: ${colors.gray500};
-  font-weight: ${fontWeight.regular};
-  font-size: ${fontSize.sm};
-`;
