+++ src/components/commons/Radio/Radio.jsx
@@ -1,8 +1,7 @@
 import React, { forwardRef } from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
-import { colors, fontSize } from "@/_shared";
+import { Layout, Icon, Label } from "./Radio.styled";
 
 const THEME = {
   LIGHT: "light",
@@ -30,44 +29,3 @@ Radio.defaultProps = {
 };
 
 export default Radio;
-
-const borderColor = {
-  light: colors.gray300,
-  dark: colors.gray600,
-};
-
-const Layout = styled.label`
-  display: flex;
-  align-content: center;
-  gap: 0.5rem;
-
-  input {
-    display: none;
-  }
-
-  input:checked + .icon {
-    border: 5px solid ${colors.blue100};
-  }
-`;
-
-const Icon = styled.div`
-  position: relative;
-  flex-shrink: 0;
-
-  width: 18px;
-  height: 18px;
-  border: 1px solid ${(props) => borderColor[props.theme]};
-  border-radius: 50%;
-
-  background-color: transparent;
-
-  cursor: pointer;
-  transition: 0.2s;
-`;
-
-const Label = styled.span`
-  color: ${colors.gray500};
-  font-size: ${fontSize.sm};
-  line-height: 18px;
-  user-select: none;
-`;
